+++ course/models.py
@@ -1,4 +1,3 @@
-from django import forms
 from django.db import models
 from slide.models import Slide
 from task.models import Task
@@ -16,25 +15,3 @@ class Course(models.Model):
 
     def __str__(self):
         return self.code + " - " + self.title
-
-
-class CourseForm(forms.ModelForm):
-
-    class Meta:
-        model = Course
-        fields = ['code', 'title', 'description', 'teacher']
-        labels = {
-            'code': 'Course code',
-            'title': 'Course name',
-            'description': 'Description',
-            'teacher': 'Teachers'
-        }
-
-
-class DeleteCourseForm(forms.Form):
-
-    confirmDelete = forms.ChoiceField(choices=((True, "Yes, delete"),
-                                               (False, "No, keep course")),
-                                      widget=forms.RadioSelect,
-                                      label=""
-                                      )
